feat: introduce AbstractModel
